Create BookInventory1.1.java
i did not want to run over the other BookInventory1 class so here is my edit it includes Nina's driver and some edits on my part. I am too sleepy to figure out if fileInventory is supposed to accept strings or file stream names. It is still untested. I wrote a displayInventory method and fixed a bunch of things including stuff in my Book class